ddms\classes\command\NewApp: Code clean up. Refactored tests\command\NewAppTest to use new tests\traits\TestsCreateApps trait to handle logic that will be common to all Test classes whose tests create a Darling Data Managment System App. This new test trait will also handle tear down so long as App's are registered via TestsCreateApps::registerApp(). This trait will make it easier to clean up any Apps that were created during tests, and also provide a single place to define methods common to Tests classes that test Commands that create Apps, or create files/directories/components for Apps. All PhpUnit tests and PhpStan tests are still passing.
